requirements.txt
Django==4.2
djangorestframework==3.14.0
pytest-django==4.5.2
pytest-asyncio==0.21.1
channels==4.0.0
channels-redis==4.1.0
websockets==11.0
pytest-mock==3.10.0
asyncio==3.4.3